backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
transformers
torch
detoxify
Pillow
dj-database-url==2.1.0